+++ price_monitor/product_advertising_api/tasks.py
@@ -28,6 +28,8 @@ from django.db.models import (
 )
 from django.utils import timezone
 
+from django.utils.translation import ugettext
+
 from price_monitor import app_settings
 from price_monitor.models import (
     Price,
@@ -311,7 +313,7 @@ class NotifySubscriberTask(Task):
 
         logger.info('Trying to send notification email to %s...', subscription.email_notification.email)
         try:
-            send_mail(product, subscription, price)
+            send_mail(product, subscription, price, self.get_audience_rating_info(product))
         except SMTPServerDisconnected:
             logger.exception('SMTP server was disconnected.')
         else:
@@ -321,3 +323,41 @@ class NotifySubscriberTask(Task):
             return True
 
         return False
+
+    def get_audience_rating_info(self, product):
+        """
+        Checks, if the product matches specific audience rating and includes additional information.
+
+        If the region is DE and the product is a FSK 18 one, additionally get all other FSK 18 products and put them into a mailable list.
+        see https://github.com/ponyriders/django-amazon-price-monitor/issues/92
+
+        As we do not currently have any use cases that could be generalized to something using the audience rating this is a country specific implementation.
+        :param product: the product to check
+        :type product:  price_monitor.models.Product
+        :return: an additional mail text or empty string if product and installation do not match prerequisites.
+        :rtype: str
+        """
+        if app_settings.PRICE_MONITOR_AMAZON_PRODUCT_API_REGION == 'DE' and product.audience_rating == 'Freigegeben ab 18 Jahren':
+            # mail text
+            mail_text = ''
+
+            # fetch all other products with FSK 18
+            for p in Product.objects.filter(audience_rating='Freigegeben ab 18 Jahren').exclude(pk=product.pk).order_by('current_price'):
+                mail_text += '{title:s}\n'.format(title=p.get_title())
+                mail_text += '{price:0.2f} {currency:s} ({price_date:s})\n'.format(
+                    price=p.current_price.value,
+                    currency=p.current_price.currency,
+                    price_date=p.current_price.date_seen.strftime('%b %d, %Y %H:%M %p %Z'),
+                )
+                mail_text += '{offer_url:s}\n'.format(offer_url=p.offer_url)
+                mail_text += '{product_detail_url:s}\n\n'.format(product_detail_url=product.get_detail_url())
+
+            # prepend introduction if there were results
+            if mail_text:
+                mail_text = '\n{intro:s}\n\n'.format(
+                    intro=ugettext('As this is a FSK 18 article, here are your other subscribed FSK 18 articles:')
+                ) + mail_text
+
+            # return
+            return mail_text
+        return ''